extract submission generate missing labels steps into a file
