+++ lib/ontologies_linked_data/models/ontology_submission.rb
@@ -608,303 +608,6 @@ module LinkedData
       end
 
 
-      def process_callbacks(logger, callbacks, action_name, &block)
-        callbacks.delete_if do |_, callback|
-          begin
-            if callback[action_name]
-              callable = self.method(callback[action_name])
-              yield(callable, callback)
-            end
-            false
-          rescue Exception => e
-            logger.error("#{e.class}: #{e.message}\n#{e.backtrace.join("\n\t")}")
-            logger.flush
-
-            if callback[:status]
-              add_submission_status(callback[:status].get_error_status)
-              self.save
-            end
-
-            # halt the entire processing if :required is set to true
-            raise e if callback[:required]
-            # continue processing of other callbacks, but not this one
-            true
-          end
-        end
-      end
-
-      def loop_classes(logger, raw_paging, callbacks)
-        page = 1
-        size = 2500
-        count_classes = 0
-        acr = self.id.to_s.split("/")[-1]
-        operations = callbacks.values.map { |v| v[:op_name] }.join(", ")
-
-        time = Benchmark.realtime do
-          paging = raw_paging.page(page, size)
-          cls_count_set = false
-          cls_count = class_count(logger)
-
-          if cls_count > -1
-            # prevent a COUNT SPARQL query if possible
-            paging.page_count_set(cls_count)
-            cls_count_set = true
-          else
-            cls_count = 0
-          end
-
-          iterate_classes = false
-          # 1. init artifacts hash if not explicitly passed in the callback
-          # 2. determine if class-level iteration is required
-          callbacks.each { |_, callback| callback[:artifacts] ||= {};
-          if callback[:caller_on_each]
-            iterate_classes = true
-          end }
-
-          process_callbacks(logger, callbacks, :caller_on_pre) {
-            |callable, callback| callable.call(callback[:artifacts], logger, paging) }
-
-          page_len = -1
-          prev_page_len = -1
-
-          begin
-            t0 = Time.now
-            page_classes = paging.page(page, size).all
-            total_pages = page_classes.total_pages
-            page_len = page_classes.length
-
-            # nothing retrieved even though we're expecting more records
-            if total_pages > 0 && page_classes.empty? && (prev_page_len == -1 || prev_page_len == size)
-              j = 0
-              num_calls = LinkedData.settings.num_retries_4store
-
-              while page_classes.empty? && j < num_calls do
-                j += 1
-                logger.error("Empty page encountered. Retrying #{j} times...")
-                sleep(2)
-                page_classes = paging.page(page, size).all
-                unless page_classes.empty?
-                  logger.info("Success retrieving a page of #{page_classes.length} classes after retrying #{j} times...")
-                end
-              end
-
-              if page_classes.empty?
-                msg = "Empty page #{page} of #{total_pages} persisted after retrying #{j} times. #{operations} of #{acr} aborted..."
-                logger.error(msg)
-                raise msg
-              end
-            end
-
-            if page_classes.empty?
-              if total_pages > 0
-                logger.info("The number of pages reported for #{acr} - #{total_pages} is higher than expected #{page - 1}. Completing #{operations}...")
-              else
-                logger.info("Ontology #{acr} contains #{total_pages} pages...")
-              end
-              break
-            end
-
-            prev_page_len = page_len
-            logger.info("#{acr}: page #{page} of #{total_pages} - #{page_len} ontology terms retrieved in #{Time.now - t0} sec.")
-            logger.flush
-            count_classes += page_classes.length
-
-            process_callbacks(logger, callbacks, :caller_on_pre_page) {
-              |callable, callback| callable.call(callback[:artifacts], logger, paging, page_classes, page) }
-
-            page_classes.each { |c|
-              # For real this is calling "generate_missing_labels_each". Is it that hard to be clear in your code?
-              # It is unreadable, not stable and not powerful. What did you want to do?
-              process_callbacks(logger, callbacks, :caller_on_each) {
-                |callable, callback| callable.call(callback[:artifacts], logger, paging, page_classes, page, c) }
-            } if iterate_classes
-
-            process_callbacks(logger, callbacks, :caller_on_post_page) {
-              |callable, callback| callable.call(callback[:artifacts], logger, paging, page_classes, page) }
-            cls_count += page_classes.length unless cls_count_set
-
-            page = page_classes.next? ? page + 1 : nil
-          end while !page.nil?
-
-          callbacks.each { |_, callback| callback[:artifacts][:count_classes] = cls_count }
-          process_callbacks(logger, callbacks, :caller_on_post) {
-            |callable, callback| callable.call(callback[:artifacts], logger, paging) }
-        end
-
-        logger.info("Completed #{operations}: #{acr} in #{time} sec. #{count_classes} classes.")
-        logger.flush
-
-        # set the status on actions that have completed successfully
-        callbacks.each do |_, callback|
-          if callback[:status]
-            add_submission_status(callback[:status])
-            self.save
-          end
-        end
-      end
-
-      def generate_missing_labels_pre(artifacts = {}, logger, paging)
-        file_path = artifacts[:file_path]
-        artifacts[:save_in_file] = File.join(File.dirname(file_path), "labels.ttl")
-        artifacts[:save_in_file_mappings] = File.join(File.dirname(file_path), "mappings.ttl")
-        property_triples = LinkedData::Utils::Triples.rdf_for_custom_properties(self)
-        Goo.sparql_data_client.append_triples(self.id, property_triples, mime_type = "application/x-turtle")
-        fsave = File.open(artifacts[:save_in_file], "w")
-        fsave.write(property_triples)
-        fsave_mappings = File.open(artifacts[:save_in_file_mappings], "w")
-        artifacts[:fsave] = fsave
-        artifacts[:fsave_mappings] = fsave_mappings
-      end
-
-      def generate_missing_labels_pre_page(artifacts = {}, logger, paging, page_classes, page)
-        artifacts[:label_triples] = []
-        artifacts[:mapping_triples] = []
-      end
-
-      # Generate labels when no label found in the prefLabel attribute (it checks rdfs:label and take label from the URI if nothing else found)
-      def generate_missing_labels_each(artifacts = {}, logger, paging, page_classes, page, c)
-        prefLabel = nil
-
-        if c.prefLabel.nil?
-          begin
-            # in case there is no skos:prefLabel or rdfs:label from our main_lang
-            rdfs_labels = c.label
-
-            if rdfs_labels && rdfs_labels.length > 1 && c.synonym.length > 0
-              rdfs_labels = (Set.new(c.label) - Set.new(c.synonym)).to_a.first
-
-              rdfs_labels = c.label if rdfs_labels.nil? || rdfs_labels.length == 0
-            end
-
-            rdfs_labels = [rdfs_labels] if rdfs_labels and not (rdfs_labels.instance_of? Array)
-            label = nil
-
-            if rdfs_labels && rdfs_labels.length > 0
-              label = rdfs_labels[0]
-            else
-              # If no label found, we take the last fragment of the URI
-              label = LinkedData::Utils::Triples.last_iri_fragment c.id.to_s
-            end
-          rescue Goo::Base::AttributeNotLoaded => e
-            label = LinkedData::Utils::Triples.last_iri_fragment c.id.to_s
-          end
-          artifacts[:label_triples] << LinkedData::Utils::Triples.label_for_class_triple(
-            c.id, Goo.vocabulary(:metadata_def)[:prefLabel], label)
-          prefLabel = label
-        else
-          prefLabel = c.prefLabel
-        end
-
-        if self.ontology.viewOf.nil?
-          loomLabel = OntologySubmission.loom_transform_literal(prefLabel.to_s)
-
-          if loomLabel.length > 2
-            artifacts[:mapping_triples] << LinkedData::Utils::Triples.loom_mapping_triple(
-              c.id, Goo.vocabulary(:metadata_def)[:mappingLoom], loomLabel)
-          end
-          artifacts[:mapping_triples] << LinkedData::Utils::Triples.uri_mapping_triple(
-            c.id, Goo.vocabulary(:metadata_def)[:mappingSameURI], c.id)
-        end
-      end
-
-      def generate_missing_labels_post_page(artifacts = {}, logger, paging, page_classes, page)
-        rest_mappings = LinkedData::Mappings.migrate_rest_mappings(self.ontology.acronym)
-        artifacts[:mapping_triples].concat(rest_mappings)
-
-        if artifacts[:label_triples].length > 0
-          logger.info("Asserting #{artifacts[:label_triples].length} labels in " +
-                        "#{self.id.to_ntriples}")
-          logger.flush
-          artifacts[:label_triples] = artifacts[:label_triples].join("\n")
-          artifacts[:fsave].write(artifacts[:label_triples])
-          t0 = Time.now
-          Goo.sparql_data_client.append_triples(self.id, artifacts[:label_triples], mime_type = "application/x-turtle")
-          t1 = Time.now
-          logger.info("Labels asserted in #{t1 - t0} sec.")
-          logger.flush
-        else
-          logger.info("No labels generated in page #{page}.")
-          logger.flush
-        end
-
-        if artifacts[:mapping_triples].length > 0
-          logger.info("Asserting #{artifacts[:mapping_triples].length} mappings in " +
-                        "#{self.id.to_ntriples}")
-          logger.flush
-          artifacts[:mapping_triples] = artifacts[:mapping_triples].join("\n")
-          artifacts[:fsave_mappings].write(artifacts[:mapping_triples])
-
-          t0 = Time.now
-          Goo.sparql_data_client.append_triples(self.id, artifacts[:mapping_triples], mime_type = "application/x-turtle")
-          t1 = Time.now
-          logger.info("Mapping labels asserted in #{t1 - t0} sec.")
-          logger.flush
-        end
-      end
-
-      def generate_missing_labels_post(artifacts = {}, logger, paging)
-        logger.info("end generate_missing_labels traversed #{artifacts[:count_classes]} classes")
-        logger.info("Saved generated labels in #{artifacts[:save_in_file]}")
-        artifacts[:fsave].close()
-        artifacts[:fsave_mappings].close()
-        logger.flush
-      end
-
-      def generate_obsolete_classes(logger, file_path)
-        self.bring(:obsoleteProperty) if self.bring?(:obsoleteProperty)
-        self.bring(:obsoleteParent) if self.bring?(:obsoleteParent)
-        classes_deprecated = []
-        if self.obsoleteProperty &&
-          self.obsoleteProperty.to_s != "http://www.w3.org/2002/07/owl#deprecated"
-
-          predicate_obsolete = RDF::URI.new(self.obsoleteProperty.to_s)
-          query_obsolete_predicate = <<eos
-SELECT ?class_id ?deprecated
-FROM #{self.id.to_ntriples}
-WHERE { ?class_id #{predicate_obsolete.to_ntriples} ?deprecated . }
-eos
-          Goo.sparql_query_client.query(query_obsolete_predicate).each_solution do |sol|
-            classes_deprecated << sol[:class_id].to_s unless ["0", "false"].include? sol[:deprecated].to_s
-          end
-          logger.info("Obsolete found #{classes_deprecated.length} for property #{self.obsoleteProperty.to_s}")
-        end
-        if self.obsoleteParent.nil?
-          # try to find oboInOWL obsolete.
-          obo_in_owl_obsolete_class = LinkedData::Models::Class
-                                        .find(LinkedData::Utils::Triples.obo_in_owl_obsolete_uri)
-                                        .in(self).first
-          self.obsoleteParent = LinkedData::Utils::Triples.obo_in_owl_obsolete_uri if obo_in_owl_obsolete_class
-        end
-        if self.obsoleteParent
-          class_obsolete_parent = LinkedData::Models::Class
-                                    .find(self.obsoleteParent)
-                                    .in(self).first
-          if class_obsolete_parent
-            descendents_obsolete = class_obsolete_parent.descendants
-            logger.info("Found #{descendents_obsolete.length} descendents of obsolete root #{self.obsoleteParent.to_s}")
-            descendents_obsolete.each do |obs|
-              classes_deprecated << obs.id
-            end
-          else
-            logger.error("Submission #{self.id.to_s} obsoleteParent #{self.obsoleteParent.to_s} not found")
-          end
-        end
-        if classes_deprecated.length > 0
-          classes_deprecated.uniq!
-          logger.info("Asserting owl:deprecated statement for #{classes_deprecated} classes")
-          save_in_file = File.join(File.dirname(file_path), "obsolete.ttl")
-          fsave = File.open(save_in_file, "w")
-          classes_deprecated.each do |class_id|
-            fsave.write(LinkedData::Utils::Triples.obselete_class_triple(class_id) + "\n")
-          end
-          fsave.close()
-          result = Goo.sparql_data_client.append_triples_from_file(
-            self.id,
-            save_in_file,
-            mime_type = "application/x-turtle")
-        end
-      end
-
       def add_submission_status(status)
         valid = status.is_a?(LinkedData::Models::SubmissionStatus)
         raise ArgumentError, "The status being added is not SubmissionStatus object" unless valid
